feat: support streaming (#3)
This PR adds support for the `stream=True` kwarg for OpenAI's SDK. 

The request/response/error serializer pattern changed to a "hook"
pattern, where each function is responsible for translating the data to
our event types _and_ sending the event to the queue. This simplifies
`wrap` so that only hooks need to run around the wrapped function call.

Right now, there's a standard implementation of collecting content from
the first choice in the stream. Unfortunately we can't get `usage` so
that's lacking here.

We send 3 events on a stream:
1. `AIRequestStart` as usual
2. `AIRequestStartStream` which is useful for first byte latency
tracking
3. `AIRequestFinish` which is fired after we see `finish_reason` while
parsing SSEs