feat: add `event_context_hook` config (#6)
This PR introduces the ability to annotate events with user-defined
context. If a reserved word is used, e.g. `environment` then we will
overwrite our metadata with that and remove it from context. This way
`context` is only user-defined values.

If the hook raises an exception, we log that via `logger.exception` and
do not update event data. However, we will continue to send the event
along.

To use this, you can pass `event_context_hook` in `init` or in the
`FlaskRosnik` constructor. For example:

```
def _my_custom_context():
  return {
    "environment": "staging",
    "origin": request.headers.get("Origin")
  }

rosnik.init(event_context_hook=_my_custom_context)
```