feat: add support for OpenAI v1 SDK (#17)
OpenAI released an official v1 SDK 🎉 

To support this, we added a new provider which allows for keeping
backwards compatibility of pre-v1 SDKs.

Clients are instantiated as objects (e.g. `OpenAI`) instead of at the
module level. While that is supported in their SDK, we are not yet
supporting that in our SDK as it's primarily meant for REPLS + notebooks
and not app code.

Hooks now expect an `instance` kwarg which allows us to pull client
metadata, similar to what we were doing before off the `openai` module.